+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * Fix incorrect NSNotification name for sale registration - erikdstock
 * Add secondary outline button style - maxim
+* Adds `inputRef` to the styled `<Input>` component - yuki24
 
 ### 1.5.8